fixed waiting screen
